TDK-7271: Test Script for lpa_hal Reason for change: changes in the test script of lpa_hal Risks: Low Test Procedure: Check the ticket. Signed-off-by:vidyasree<vidyasree.g@tataelxsi.co.in>
