+++ src/test_register.c
@@ -1,0 +1,29 @@
+/*
+* If not stated otherwise in this file or this component's LICENSE file the
+* following copyright and licenses apply:*
+* Copyright 2023 RDK Management
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+/* L1 Testing Functions */
+extern int test_lpa_hal_l1_register(void);
+ 
+int register_hal_l1_tests( void )
+{
+    int registerFailed=0;
+
+    registerFailed |= test_lpa_hal_l1_register();
+ 
+    return registerFailed;
+}
